Fix: 지도 로딩 로직 분리 및 데이터 로딩 문제 해결
