+++ components/pages/map/Tab.tsx
@@ -33,8 +33,6 @@ const Tab = () => {
     }
   };
 
-  console.log(searchResult);
-
   return (
     <>
       {tab ? (
