+++ tailwind.config.ts
@@ -1,30 +1,30 @@
-import type { Config } from "tailwindcss";
+import type { Config } from 'tailwindcss';
 
 const createPxEntries = (size: number) => {
   return {
-    0: "0",
+    0: '0',
     ...Array.from(Array(size + 1)).reduce((accumulator, _, i) => {
-      return { ...accumulator, [`${i * 4}`]: `${(i * 4)}px` };
+      return { ...accumulator, [`${i * 4}`]: `${i * 4}px` };
     }),
   };
 };
 
 const PX_ENTRIES = createPxEntries(500);
 
-
 const config: Config = {
   content: [
-    "./pages/**/*.{js,ts,jsx,tsx,mdx}",
-    "./components/**/*.{js,ts,jsx,tsx,mdx}",
-    "./app/**/*.{js,ts,jsx,tsx,mdx}",
+    './pages/**/*.{js,ts,jsx,tsx,mdx}',
+    './components/**/*.{js,ts,jsx,tsx,mdx}',
+    './hooks/**/*.{js,ts,jsx,tsx,mdx}',
+    './app/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   theme: {
     spacing: PX_ENTRIES,
     zIndex: {
-      base: "1",
-      nav: "2",
-      popup: "999",
-      floating: "1000",
+      base: '1',
+      nav: '2',
+      popup: '999',
+      floating: '1000',
     },
   },
   plugins: [],
